refactor: rename component folder
