+++ src/app/components/board.ts
