+++ src/app/components/digits.ts
